[kyg] 숫자 비교

+++ lec02_variable/src/com/gn/homework/Homework05.java
@@ -1,18 +1,30 @@
 package com.gn.homework;
+
 import java.util.Scanner;
+
 public class Homework05 {
 
+
+
 	public static void main(String[] args) {
+
 		Scanner sc = new Scanner(System.in);
+
 		System.out.print("국어: " );
-		int a = sc.nextInt();
+
+		double a = sc.nextDouble();
+
 		System.out.print("영어: " );
-		int b = sc.nextInt();
+
+		double b = sc.nextDouble();
+
 		System.out.print("수학: " );
-		int c = sc.nextInt();
-		System.out.println("총점: " + (a+b+c) );
-		System.out.println("평균: " + (a+b+c) / 3);
-		
+
+		double c = sc.nextDouble();
+
+		System.out.println("총점: " + (int)(a+b+c) );
+
+		System.out.println("평균: " + (int)((a+b+c) / 3));
 
 	}
